# public/openapi.yaml
openapi: 3.1.0
info:
  title: Insight Oracle Risk & Transparency API
  version: v1
  description: |
    Hourly oracle price data, cross-oracle consensus, deviation signals, feed health tracking,
    risk metrics, protocol liquidation analysis, stablecoin depeg detection, and daily risk reports.

    ---

    ## Quick Start

    1. **Get an API Key** — Create one in [Settings](/settings). It's shown only once, so store it securely.
    2. **Build the Query** — Choose a provider, symbol, and chain. URL-encode symbols like `BTC%2FUSD`.
    3. **Handle the Response** — Check the `success` flag, read `price` and `verification` metadata, handle errors.

    ```bash
    curl -H "X-API-Key: ins_your_key_here" \
      "https://oracleinsight.xyz/api/v1/prices?provider=chainlink&symbol=BTC%2FUSD&chain=ethereum"
    ```

    ---

    ## Authentication

    Most endpoints require authentication via one of two methods:

    | Method | Header | Use Case |
    |--------|--------|----------|
    | **API Key** | `X-API-Key: ins_...` | External integrations, scripts, bots |
    | **Bearer Token** | `Authorization: Bearer <jwt>` | Web UI session (Supabase Auth) |

    - Both methods are accepted on most endpoints. The `alerts` endpoints only accept Bearer tokens.
    - Public endpoints (`/health`, `/symbols`) require no authentication.

    ⚠️ **Keep your key safe** — API keys are shown only once when created. Rotate them immediately if leaked.

    ---

    ## Plans & Pricing

...

...

    ---

...

...

...

...

...

...

...
servers:
...
security: []
paths:
...
  /price-records/export:
    get:
      operationId: exportPriceRecords
      tags: [Data]
      summary: Export price records 🔒
      description: |
        Export raw price records for external analysis. Supports filtering by symbol, provider, and date range.

        **Requires Pro plan or higher.**
      security:
        - apiKey: []
        - bearer: []
      parameters:
        - name: symbol
          in: query
          schema:
            type: string
        - name: provider
          in: query
          schema:
            $ref: '#/components/schemas/OracleProvider'
        - $ref: '#/components/parameters/dateFrom'
        - $ref: '#/components/parameters/dateTo'
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50000
            default: 1000
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Price records
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceExportResponse'
        '402':
          $ref: '#/components/responses/PlanUpgradeRequired'
...
components:
...
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: 'API key with `ins_` prefix (e.g. ins_a3f1...)'
    bearer:
      type: http
      scheme: bearer
      description: 'Supabase Auth JWT session token'

  # ── Reusable Parameters ─────────────────────────────────
  parameters:
...
    dateFrom:
      name: from
      in: query
      description: 'Start date in YYYY-MM-DD format. Defaults to 7 days ago.'
      schema:
        type: string
        pattern: '^[0-9]{4}-[0-9]{2}-[0-9]{2}$'
    dateTo:
      name: to
      in: query
      description: 'End date in YYYY-MM-DD format. Defaults to today.'
      schema:
        type: string
        pattern: '^[0-9]{4}-[0-9]{2}-[0-9]{2}$'
...
  responses:
...
    PlanUpgradeRequired:
      description: Free-tier daily quota for highest-value endpoints exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            success: false
            error:
              code: PLAN_UPGRADE_REQUIRED
              message: 'Free-tier daily quota for highest-value endpoints exceeded. Upgrade to continue.'
              retryable: false
...
  schemas:
...
    OracleProvider:
      type: string
      enum:
        - chainlink
        - pyth
        - api3
        - redstone
        - dia
        - winklink
        - supra
        - twap
        - reflector
        - flare
...
    ApiMeta:
      type: object
      properties:
        timestamp:
          type: integer
          description: Unix timestamp in milliseconds
        requestId:
          type: string
          description: Unique request identifier
      required: [timestamp]
    ApiError:
      type: object
      properties:
        code:
          type: string
          description: Machine-readable error code
        message:
          type: string
          description: Human-readable error message
        retryable:
          type: boolean
        details:
          type: object
          additionalProperties: true
      required: [code, message, retryable]
    ApiErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          $ref: '#/components/schemas/ApiError'
        meta:
          $ref: '#/components/schemas/ApiMeta'
      required: [success, error]
    OnChainVerification:
      type: object
      properties:
        type:
          type: string
          enum: [on-chain, api]
        contractAddress:
          type: string
        chainId:
          type: integer
        explorerUrl:
          type: string
          description: Link to blockchain explorer contract overview page
        method:
          type: string
          description: The on-chain method used to read the price (e.g. latestRoundData)
        blockNumber:
          type: integer
...
    PriceExportRecord:
      type: object
      properties:
        provider:
          type: string
        symbol:
          type: string
        chain:
          type: string
        price:
          type: number
        timestamp:
          type: integer
        confidence:
          type: number
        source:
          type: string
        verification:
          $ref: '#/components/schemas/OnChainVerification'
    PriceExportResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            exportedAt:
              type: string
            filters:
              type: object
              properties:
                symbol:
                  type: string
                provider:
                  type: string
                from:
                  type: string
                to:
                  type: string
            count:
              type: integer
            records:
              type: array
              items:
                $ref: '#/components/schemas/PriceExportRecord'
        meta:
          $ref: '#/components/schemas/ApiMeta'

    # ── Daily Report ──────────────────────────────────
...